v1.2
It is now possible to save and load databases using the pickle (de)serialization library. These files can be loaded into main.py by dragging them onto it, which works on it's own to continue crawling. A new load_graph script allows visualizing graphs from the .webdb files.

Another big feature is the ability to use pyVis as a renderer instead of the default NetworkX visualization.

A number of small improvements have also been made to the overall code structure.